requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Pillow==8.1.2
pip==21.0.1
requests==2.25.1
setuptools==54.2.0
urllib3==1.26.4
